monitors/RES/source/requirements.txt
psutil == 5.8.0          
pyyaml == 5.3.1
pika == 1.2.0
